With 1.4 toggle applied.
Additional test for that.

+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/task/RespondToClaimDraftDeletionScheduledTask.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.pcs.ccd.service.respondpossessionclaim.DraftResponseDeletionService;
+import uk.gov.hmcts.reform.pcs.service.FeatureFlag;
+import uk.gov.hmcts.reform.pcs.service.FeatureToggleService;
 
 import java.time.Duration;
 
@@ -24,18 +26,21 @@ public class RespondToClaimDraftDeletionScheduledTask {
     private final Duration backoffDelay;
 
     private final DraftResponseDeletionService draftResponseDeletionService;
+    private final FeatureToggleService featureToggleService;
 
     public RespondToClaimDraftDeletionScheduledTask(
             @Value("${respond-to-claim-draft-deletion.schedule}") String schedule,
             @Value("${respond-to-claim-draft-deletion.discard-after-days}") int discardAfterDays,
             @Value("${respond-to-claim-draft-deletion.request.max-retries:3}") int maxRetries,
             @Value("${respond-to-claim-draft-deletion.request.backoff-delay-seconds:10}") Duration backoffDelay,
-            DraftResponseDeletionService draftResponseDeletionService) {
+            DraftResponseDeletionService draftResponseDeletionService,
+            FeatureToggleService featureToggleService) {
         this.schedule = schedule;
         this.discardAfterDays = discardAfterDays;
         this.maxRetries = maxRetries;
         this.backoffDelay = backoffDelay;
         this.draftResponseDeletionService = draftResponseDeletionService;
+        this.featureToggleService = featureToggleService;
     }
 
     @Bean
@@ -49,9 +54,13 @@ public class RespondToClaimDraftDeletionScheduledTask {
     }
 
     public void runSweep() {
-        log.info("runSweep starting up ...");
-        draftResponseDeletionService.deleteRespondPossessionClaimBatch(discardAfterDays);
-        log.info("--- runSweep closing down");
+        if (featureToggleService.isEnabled(FeatureFlag.RELEASE_1_DOT_4)) {
+            log.info("runSweep starting up ...");
+            draftResponseDeletionService.deleteRespondPossessionClaimBatch(discardAfterDays);
+            log.info("--- runSweep closing down");
+        } else {
+            log.info("Not enabled in this release.");
+        }
     }
 
 }
